Extract abstract class for TypeScript generators
